refactor(server): relocate session-lifecycle handlers
Move handleCancelCurrentTurn, handleStopSession, handleClearSession, the
coding-CLI tmux close helpers (closeAllCodingCLIInteractiveSessionsForOwner,
gracefulCloseCodingCLITmuxByName), and the browser-session listing/cleanup
handlers (~393 lines) out of server.go into session_lifecycle.go. Pure
relocation: same package, no behavior change.

server.go now 8,505 lines (down from 11,738 across all relocations).
fmt/vet clean, cmd/server tests pass.
