+++ agent_go/cmd/server/server.go
@@ -5488,400 +5488,6 @@ func (api *StreamingAPI) verifySessionAccess(r *http.Request, sessionID string) 
 	return nil
 }
 
-// handleCancelCurrentTurn cancels only the currently running LLM turn for a session.
-// It must not mark the session stopped or tear down workshop/background state.
-func (api *StreamingAPI) handleCancelCurrentTurn(w http.ResponseWriter, r *http.Request) {
-	sessionID := r.Header.Get("X-Session-ID")
-	if sessionID == "" {
-		http.Error(w, "Session ID required", http.StatusBadRequest)
-		return
-	}
-
-	if err := api.verifySessionAccess(r, sessionID); err != nil {
-		http.Error(w, err.Error(), http.StatusNotFound)
-		return
-	}
-
-	api.agentCancelMux.Lock()
-	cancelFunc, exists := api.agentCancelFuncs[sessionID]
-	if exists {
-		cancelFunc()
-		delete(api.agentCancelFuncs, sessionID)
-		log.Printf("[SESSION DEBUG] Canceled current LLM turn for session %s", sessionID)
-	}
-	api.agentCancelMux.Unlock()
-
-	if !exists {
-		log.Printf("[SESSION DEBUG] No active LLM turn to cancel for session %s", sessionID)
-	}
-
-	w.WriteHeader(http.StatusNoContent)
-}
-
-// closeAllCodingCLIInteractiveSessionsForOwner tears down the persistent
-// tmux-backed coding-CLI session registered under the given owner key, across
-// every tmux provider. Each adapter's CloseXxxInteractiveSessionForOwner runs
-// its own graceful-then-force shutdown (e.g. agy: tmux send-keys "/exit" →
-// tmux kill-session) and is a no-op when no session is registered for the
-// owner, so calling all five is safe and provider-agnostic. (OpenCode is a
-// structured transport with no persistent tmux session, so it needs no close.)
-func closeAllCodingCLIInteractiveSessionsForOwner(owner, reason string) {
-	llmproviders.CloseAgyCLIInteractiveSessionForOwner(owner, reason)
-	llmproviders.CloseCursorCLIInteractiveSessionForOwner(owner, reason)
-	llmproviders.CloseGeminiCLIInteractiveSessionForOwner(owner, reason)
-	llmproviders.CloseCodexCLIInteractiveSessionForOwner(owner, reason)
-	llmproviders.CloseClaudeCodeInteractiveSessionForOwner(owner, reason)
-}
-
-// gracefulCloseCodingCLITmuxByName runs the provider-specific graceful shutdown
-// (e.g. agy: Escape → "/exit" → Enter; claude: "C-u /exit C-m"; codex/cursor/
-// gemini: C-c) plus the adapter's file/MCP-lease cleanup for the tmux-backed
-// coding CLI identified by its tmux session name. The provider is detected from
-// the session-name prefix (set by each adapter's new<Provider>TmuxSessionName).
-// This tears the session down by tmux name rather than owner key, so it works
-// for workflow sub-agents that registered under a step-execution owner the
-// caller can't reconstruct. Returns false when the prefix matches no known
-// provider (caller should fall back to a raw kill-session).
-func gracefulCloseCodingCLITmuxByName(tmuxName, reason string) bool {
-	name := strings.TrimSpace(tmuxName)
-	if name == "" {
-		return false
-	}
-	switch {
-	case strings.HasPrefix(name, "mlp-agy-cli"):
-		llmproviders.CloseAgyCLIInteractiveSessionByTmux(name, reason)
-	case strings.HasPrefix(name, "mlp-claude-code"):
-		llmproviders.CloseClaudeCodeInteractiveSessionByTmux(name, reason)
-	case strings.HasPrefix(name, "mlp-codex-cli"):
-		llmproviders.CloseCodexCLIInteractiveSessionByTmux(name, reason)
-	case strings.HasPrefix(name, "mlp-cursor-cli"):
-		llmproviders.CloseCursorCLIInteractiveSessionByTmux(name, reason)
-	case strings.HasPrefix(name, "mlp-gemini-cli"):
-		llmproviders.CloseGeminiCLIInteractiveSessionByTmux(name, reason)
-	default:
-		return false
-	}
-	return true
-}
-
-// Add endpoint to stop/clear a session
-func (api *StreamingAPI) handleStopSession(w http.ResponseWriter, r *http.Request) {
-	sessionID := r.Header.Get("X-Session-ID")
-	if sessionID == "" {
-		http.Error(w, "Session ID required", http.StatusBadRequest)
-		return
-	}
-
-	if err := api.verifySessionAccess(r, sessionID); err != nil {
-		http.Error(w, err.Error(), http.StatusNotFound)
-		return
-	}
-
-	// Mark session as stopped FIRST, before any cancellation, so that in-flight
-	// goroutines that race with this stop handler will see the flag and bail out
-	// instead of re-creating workshop sessions or spawning new CLI processes.
-	// See stoppedSessions field comment for the full race condition description.
-	api.markSessionStopped(sessionID)
-
-	// Cancel agent execution context if it exists
-	api.agentCancelMux.Lock()
-	if cancelFunc, exists := api.agentCancelFuncs[sessionID]; exists {
-		cancelFunc() // Cancel the agent execution
-		delete(api.agentCancelFuncs, sessionID)
-		log.Printf("[SESSION DEBUG] Canceled agent execution context for session %s", sessionID)
-	}
-	api.agentCancelMux.Unlock()
-
-	// Update active session status to stopped
-	api.updateSessionStatus(sessionID, "stopped")
-	api.setSessionBusy(sessionID, false)
-	api.setSyntheticTurn(sessionID, false)
-
-	// Cancel background agents if explicitly requested (e.g. user pressed the stop button).
-	// When called before sending a new message, cancelAgents is NOT set so agents survive
-	// across turns and synthetic turns can still fire when they complete.
-	if r.URL.Query().Get("cancelAgents") == "true" {
-		api.bgAgentRegistry.CancelAll(sessionID)
-		log.Printf("[SESSION DEBUG] Canceled all background agents for session %s", sessionID)
-	}
-
-	// Prevent stopped sessions from being revived by queued background completions or
-	// synthetic auto-notification turns that reuse the stored agent after stop.
-	api.pendingMu.Lock()
-	delete(api.pendingCompletions, sessionID)
-	delete(api.completionRetryScheduled, sessionID)
-	api.pendingMu.Unlock()
-
-	// Clear pending start notifications so they don't leak across stop/restart
-	// (pending-start-notifications-leak fix).
-	api.pendingStartMu.Lock()
-	delete(api.pendingStartNotifications, sessionID)
-	delete(api.startNotificationRetryScheduled, sessionID)
-	api.pendingStartMu.Unlock()
-
-	api.lastQueryMu.Lock()
-	delete(api.lastQueryRequests, sessionID)
-	api.lastQueryMu.Unlock()
-
-	api.sessionWorkspaceMu.Lock()
-	delete(api.sessionWorkspaceFolders, sessionID)
-	api.sessionWorkspaceMu.Unlock()
-
-	api.sessionAgentsMux.Lock()
-	delete(api.sessionAgents, sessionID)
-	api.sessionAgentsMux.Unlock()
-
-	api.completionLoopStartedMu.Lock()
-	delete(api.completionLoopStarted, sessionID)
-	api.completionLoopStartedMu.Unlock()
-
-	api.bgAgentRegistry.Cleanup(sessionID)
-	log.Printf("[SESSION DEBUG] Cleared synthetic-turn state for stopped session %s", sessionID)
-
-	// Close workshop chat sessions for this session — cancels all running step executions.
-	// Workshop sessions use context.Background() so they survive agent context cancellation above;
-	// we must explicitly call Close() to cancel their step goroutines.
-	//
-	// Close() → cancelFunc() cascades to all execCtx (step goroutines) → kills Codex CLI
-	// processes via exec.CommandContext. It also calls CloseWorkshopGroupSessions() which
-	// closes MCP connections for group sessions (session-group-*) and isolated sub-sessions.
-	//
-	// IMPORTANT: The markSessionStopped() call above prevents in-flight goroutines from
-	// re-creating the workshop after we close it here. Without that guard, a racing
-	// goroutine could call NewWorkshopChatSession() with a fresh context.Background(),
-	// creating orphaned CLI processes that are never canceled. See stoppedSessions comment.
-	//
-	// Historically we keyed this map by sessionID / "eval-"+sessionID, but some workflow
-	// execution paths can drift from those exact keys. So first try direct keys, then scan
-	// for any workshop session whose owning mainSessionID matches this session.
-	closedWorkshopKeys := map[string]struct{}{}
-	workshopKeys := []string{sessionID, "eval-" + sessionID}
-	for _, wsKey := range workshopKeys {
-		if cached, ok := api.workshopChatSessions.Load(wsKey); ok {
-			if ws, ok := cached.(interface{ Close() }); ok {
-				ws.Close()
-				log.Printf("[SESSION DEBUG] Closed workshop session %q (all step executions canceled)", wsKey)
-			}
-			api.workshopChatSessions.Delete(wsKey)
-			closedWorkshopKeys[wsKey] = struct{}{}
-		}
-	}
-	api.workshopChatSessions.Range(func(key, value interface{}) bool {
-		wsKey, ok := key.(string)
-		if !ok {
-			return true
-		}
-		if _, alreadyClosed := closedWorkshopKeys[wsKey]; alreadyClosed {
-			return true
-		}
-		ws, ok := value.(interface {
-			Close()
-			MainSessionID() string
-		})
-		if !ok || ws.MainSessionID() != sessionID {
-			return true
-		}
-		ws.Close()
-		api.workshopChatSessions.Delete(wsKey)
-		log.Printf("[SESSION DEBUG] Closed workshop session %q via mainSessionID match for session %s", wsKey, sessionID)
-		return true
-	})
-
-	// Cancel all workflow orchestrator contexts for this session
-	// Since we now use queryID as the key, we need to look up all queryIDs for this session
-	api.sessionQueryIDMux.Lock()
-	queryIDs := api.sessionQueryIDs[sessionID]
-	delete(api.sessionQueryIDs, sessionID) // Clear the mapping
-	api.sessionQueryIDMux.Unlock()
-
-	if len(queryIDs) > 0 {
-		// Cancel all background agents BEFORE canceling workflow contexts.
-		// When workflow contexts are canceled, sub-agent goroutines will eventually
-		// fail with "context canceled" and call OnExecutionComplete. Without marking
-		// them as canceled first, they'd fire stale AUTO-NOTIFICATION synthetic turns.
-		api.bgAgentRegistry.CancelAll(sessionID)
-		log.Printf("[SESSION DEBUG] Canceled all background agents for session %s (workflow stop)", sessionID)
-
-		api.workflowOrchestratorContextMux.Lock()
-		for _, qid := range queryIDs {
-			if cancelFunc, exists := api.workflowOrchestratorContexts[qid]; exists {
-				cancelFunc() // Cancel this workflow execution
-				delete(api.workflowOrchestratorContexts, qid)
-				log.Printf("[SESSION DEBUG] Canceled workflow execution %s for session %s", qid, sessionID)
-			}
-		}
-		api.workflowOrchestratorContextMux.Unlock()
-
-		// Remove from active executions registry
-		api.activeWorkflowExecutionsMux.Lock()
-		for _, qid := range queryIDs {
-			delete(api.activeWorkflowExecutions, qid)
-		}
-		api.activeWorkflowExecutionsMux.Unlock()
-		api.cancelTrackedExecutionsForSession(sessionID)
-		log.Printf("[SESSION DEBUG] Canceled %d workflow execution(s) for session %s", len(queryIDs), sessionID)
-	}
-
-	// Clear workflow objective
-	api.workflowObjectiveMux.Lock()
-	if _, exists := api.workflowObjectives[sessionID]; exists {
-		delete(api.workflowObjectives, sessionID)
-		log.Printf("[SESSION DEBUG] Cleared workflow objective for session %s", sessionID)
-	}
-	api.workflowObjectiveMux.Unlock()
-
-	// Close all MCP sessions (browsers, etc.) associated with this HTTP session immediately.
-	// This is safe to call even if the defers in the workflow goroutines haven't fired yet —
-	// CloseSession is idempotent, so those defers will be no-ops when they eventually run.
-	log.Printf("[SESSION DEBUG] Closing MCP sessions for stopped session %s", sessionID)
-	mcpagent.CloseHTTPSession(sessionID)
-
-	// Kill headless browser processes for this session
-	api.cleanupBrowserSessions(sessionID)
-
-	// Close any tmux-backed coding-CLI session for this chat. Without this,
-	// canceling the Go context above tears down the streaming connection
-	// server-side but the CLI process inside the tmux pane keeps running
-	// its current turn (LLM calls, shell commands) until it finishes
-	// naturally — the user pressed stop but agy/codex/etc. ran for another
-	// 30-60 seconds.
-	//
-	// Each adapter's CloseXxxInteractiveSessionForOwner implements that
-	// CLI's graceful-then-force shutdown sequence (agy: tmux send-keys
-	// "/exit" → tmux kill-session; codex/cursor/gemini/claude-code:
-	// adapter-specific exit → tmux kill-session). All are no-ops when no
-	// session is registered for the owner, so calling all five is safe
-	// and provider-agnostic.
-	closeReason := "user pressed stop"
-	closeAllCodingCLIInteractiveSessionsForOwner(sessionID, closeReason)
-	log.Printf("[SESSION DEBUG] Closed any tmux-backed coding-CLI session for stopped session %s", sessionID)
-
-	// The calls above are keyed by the chat / main-agent session ID. Workflow-step
-	// sub-agents, however, register their interactive CLI session under the STEP
-	// execution-owner ID (e.g. "workflow-step:exec-...:step-name") — not this
-	// chat sessionID — so the owner-keyed close above never matches them and
-	// their tmux panes orphan: the CLI process keeps running (and keeps holding
-	// the workflow's single-session slot, which blocks "new chat" and workflow
-	// model changes) long after the user stopped the run. This is provider-wide,
-	// not agy-specific.
-	//
-	// When the caller explicitly asked to cancel agents (cancelAgents=true — e.g.
-	// the workflow "kill & start new chat" popup, which calls stopSession(sid,
-	// true)), enumerate this session's live terminals and tear each sub-agent
-	// down by its OWN owner key (graceful, provider-agnostic), plus a guaranteed
-	// tmux kill-session backstop for active panes in case the adapter's registry
-	// bookkeeping drifted from the terminal store's owner ID.
-	if r.URL.Query().Get("cancelAgents") == "true" && api.terminalStore != nil {
-		mainOwner := "main:" + sessionID
-		for _, snap := range api.terminalStore.List(sessionID) {
-			owner := strings.TrimSpace(snap.OwnerID)
-			if owner == "" || owner == sessionID || owner == mainOwner {
-				continue // chat / main-agent session already handled above
-			}
-			tmux := strings.TrimSpace(snap.TmuxSession)
-			if tmux == "" {
-				continue // structured transport: no tmux pane; context cancel above handles it
-			}
-			// Primary: run the provider's own graceful exit + cleanup, resolved
-			// by tmux name so it works even though the sub-agent registered its
-			// CLI session under a step-execution owner (not this chat sessionID).
-			// The adapter kills the tmux session itself as the final step of that
-			// sequence, so a separate kill-session is only needed when no adapter
-			// claimed it (e.g. the registry was lost across a server restart).
-			if handled := gracefulCloseCodingCLITmuxByName(tmux, closeReason); !handled {
-				killCtx, killCancel := context.WithTimeout(context.Background(), terminalTmuxActionTimeout)
-				if err := runTerminalTmuxCommand(killCtx, "", "kill-session", "-t", tmux); err != nil {
-					log.Printf("[SESSION DEBUG] kill-session %q (owner %s) failed (may already be gone): %v", tmux, owner, err)
-				}
-				killCancel()
-			}
-			if snap.Active {
-				// Only relabel panes that were still live; leave already completed
-				// terminals in their terminal state.
-				api.terminalStore.MarkFailed(snap.TerminalID)
-			}
-			log.Printf("[SESSION DEBUG] Tore down workflow sub-agent terminal owner=%s tmux=%s active=%v for stopped session %s", owner, tmux, snap.Active, sessionID)
-		}
-	}
-
-	// Note: Conversation history and orchestrator state are preserved to allow resuming the conversation
-	// Use /api/session/clear if you want to clear conversation history
-
-	w.WriteHeader(http.StatusOK)
-	w.Write([]byte("Session stopped (conversation history and orchestrator state preserved)"))
-}
-
-// handleGetBrowserSessions returns the tracked browser sessions with their owning chat session IDs.
-func (api *StreamingAPI) handleGetBrowserSessions(w http.ResponseWriter, r *http.Request) {
-	tracker := browser.GetSessionTracker()
-	sessions := tracker.ActiveSessions()
-	cdpOwners := browser.ActiveCDPOwnersSnapshot()
-	w.Header().Set("Content-Type", "application/json")
-	json.NewEncoder(w).Encode(map[string]interface{}{
-		"sessions":   sessions,
-		"count":      len(sessions),
-		"cdp_owners": cdpOwners,
-	})
-}
-
-// cleanupBrowserSessions closes all headless browser processes for a session.
-// Must be called whenever a session ends (stop, clear, workflow completion).
-func (api *StreamingAPI) cleanupBrowserSessions(sessionID string) {
-	tracker := browser.GetSessionTracker()
-	if tracker.CountForChat(sessionID) == 0 {
-		return
-	}
-	workspaceAPIURL := os.Getenv("WORKSPACE_API_URL")
-	if workspaceAPIURL == "" {
-		workspaceAPIURL = "http://127.0.0.1:8081"
-	}
-	client := browser.NewClient(workspaceAPIURL)
-	tracker.CloseAllForChat(sessionID, client)
-}
-
-// Add endpoint to clear conversation history for a session
-func (api *StreamingAPI) handleClearSession(w http.ResponseWriter, r *http.Request) {
-	sessionID := r.Header.Get("X-Session-ID")
-	if sessionID == "" {
-		http.Error(w, "Session ID required", http.StatusBadRequest)
-		return
-	}
-
-	// Verify session ownership via the in-memory active sessions map.
-	if err := api.verifySessionAccess(r, sessionID); err != nil {
-		http.Error(w, "Session not found or access denied", http.StatusNotFound)
-		return
-	}
-
-	// Clear conversation and coding-agent resume state guarded by the same
-	// mutex used by query/resume paths.
-	api.conversationMux.Lock()
-	if _, exists := api.conversationHistory[sessionID]; exists {
-		delete(api.conversationHistory, sessionID)
-		log.Printf("[SESSION DEBUG] Cleared conversation history for session %s", sessionID)
-	}
-	delete(api.lastWorkshopModeBySession, sessionID)
-	api.conversationMux.Unlock()
-
-	// Clear orchestrator state (removed - now stateless)
-
-	// Clear orchestrator instance (legacy removed)
-
-	// Clear workflow objective
-	api.workflowObjectiveMux.Lock()
-	if _, exists := api.workflowObjectives[sessionID]; exists {
-		delete(api.workflowObjectives, sessionID)
-		log.Printf("[SESSION DEBUG] Cleared workflow objective for session %s", sessionID)
-	}
-	api.workflowObjectiveMux.Unlock()
-
-	// Kill headless browser processes for this session
-	api.cleanupBrowserSessions(sessionID)
-
-	w.WriteHeader(http.StatusOK)
-	w.Write([]byte("Session cleared (conversation history and orchestrator state removed)"))
-}
-
 // State management functions removed - orchestrator is now stateless
 
 // createServerLogger creates a logger instance for the server
